feat: add and refactor joi validators
