--- validators/signin-validator.js
@@ -1,8 +1,0 @@
-const { celebrate, Joi } = require('celebrate');
-
-module.exports = celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().email().required(),
-    password: Joi.string().required(),
-  }),
-});
